refactor: Aligning Map and Parallel
