+++ lib/state-machines/state-types/Map.js
@@ -1,5 +1,4 @@
 'use strict'
-const cloneDeep = require('lodash/cloneDeep')
 const ParentBaseState = require('./Parent-base-state')
 const jp = require('jsonpath')
 const States = require('./errors')
@@ -13,30 +12,18 @@ class Map extends ParentBaseState {
     // todo: MaxConcurrency
   }
 
+  buildBranchExecutions (input, parentExecutionName) {
+    const branches = jp.value(input, this.itemsPath)
+    return branches.map((item, index) =>
+      this.makeChildExecution(this.iterator, item, index, parentExecutionName)
+    )
+  } // buildBranchExecutions
+
   async process (executionDescription, input) {
     const parentExecutionName = executionDescription.executionName
-
-    const branches = jp.value(input, this.itemsPath)
-    executionDescription.childCount = branches.length
     await this.dao.checkpoint(executionDescription)
 
-    executionDescription.childCount = branches.length
-
-    const branchExecutions = branches.map((item, index) => {
-      const branchContext = cloneDeep(item)
-      return this.executioner(
-        branchContext,
-        this.iterator,
-        {
-          branchIndex: index,
-          parentExecutionName,
-          parentStateMachineName: this.stateMachineName,
-          parentStateName: this.name,
-          sendResponse: 'COMPLETED'
-        },
-        this.options
-      )
-    })
+    const branchExecutions = this.buildBranchExecutions(input, parentExecutionName)
 
     Promise
       .all(branchExecutions)
