+++ lib/state-machines/state-types/Parallel.js
@@ -1,5 +1,4 @@
 'use strict'
-const cloneDeep = require('lodash/cloneDeep')
 const ParentBaseState = require('./Parent-base-state')
 const States = require('./errors')
 
@@ -18,28 +17,17 @@ class Parallel extends ParentBaseState {
     this.debug()
   } // constructor
 
+  buildBranchExecutions (input, parentExecutionName) {
+    return this.branches.map((stateMachineName, index) =>
+      this.makeChildExecution(stateMachineName, input, index, parentExecutionName)
+    )
+  } // buildBranchExecutions
+
   async process (executionDescription, input) {
     const parentExecutionName = executionDescription.executionName
-
-    executionDescription.childCount = this.branches.length
     await this.dao.checkpoint(executionDescription)
 
-    const branchExecutions = this.branches
-      .map((stateMachineName, index) => {
-        const branchContext = cloneDeep(input)
-        return this.executioner(
-          branchContext,
-          stateMachineName,
-          {
-            branchIndex: index,
-            parentExecutionName: parentExecutionName,
-            parentStateMachineName: this.stateMachineName,
-            parentStateName: this.name,
-            sendResponse: 'COMPLETED'
-          },
-          this.options
-        )
-      })
+    const branchExecutions = this.buildBranchExecutions(input, parentExecutionName)
 
     Promise
       .all(branchExecutions)
